+++ mayan/apps/converter/backends/literals.py
@@ -1,4 +1,0 @@
-from __future__ import unicode_literals
-
-DEFAULT_PDFTOPPM_PATH = '/usr/bin/pdftoppm'
-DEFAULT_PDFINFO_PATH = '/usr/bin/pdfinfo'